+++ packages/eyes-storybook/test/fixtures/applitools-ignore-git-merge-base.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  appName: 'Simple storybook',
+  branchName: 'branch-that-exists-for-testing-do-not-delete',
+  parentBranchName: 'master',
+  ignoreGitMergeBase: true,
+};