+++ packages/eyes-storybook/test/it/eyesStorybook.it.test.js
@@ -187,4 +187,71 @@ describe('eyesStorybook', () => {
 ✔ Done 19 stories out of 19
 `);
   });
+
+  it('sends parentBranchBaselineSavedBefore when branchName and parentBranchName are specified, and there is a merge-base time for them', async () => {
+    const {stream} = testStream();
+    const configPath = path.resolve(
+      __dirname,
+      '../fixtures/applitools-ignore-git-merge-base.config.js',
+    );
+    const defaultConfig = {waitBeforeScreenshots: 50};
+    const config = generateConfig({argv: {conf: configPath}, defaultConfig, externalConfigParams});
+
+    // this is the important part, because it's not `true` then `parentBranchBaselineSavedBefore` should be sent in `startInfo`
+    delete config.ignoreGitMergeBase;
+
+    let results = await eyesStorybook({
+      config: {
+        serverUrl,
+        storybookUrl: 'http://localhost:9001',
+        ...config,
+      },
+      logger,
+      performance,
+      timeItAsync,
+      outputStream: stream,
+    });
+    results = flatten(results.map(r => r.resultsOrErr));
+    for (const testResults of results) {
+      const sessionUrl = `${serverUrl}/api/sessions/batches/${encodeURIComponent(
+        testResults.getBatchId(),
+      )}/${encodeURIComponent(testResults.getId())}`;
+
+      const session = await fetch(sessionUrl).then(r => r.json());
+      expect(session.startInfo.parentBranchBaselineSavedBefore).to.match(
+        /\d{4}-\d{2}-\d{2}T\d{2}\:\d{2}:\d{2}\+\d{2}\:\d{2}/,
+      );
+    }
+  });
+
+  it('handles ignoreGitMergeBase', async () => {
+    const {stream} = testStream();
+    const configPath = path.resolve(
+      __dirname,
+      '../fixtures/applitools-ignore-git-merge-base.config.js',
+    );
+    const defaultConfig = {waitBeforeScreenshots: 50};
+    const config = generateConfig({argv: {conf: configPath}, defaultConfig, externalConfigParams});
+
+    let results = await eyesStorybook({
+      config: {
+        serverUrl,
+        storybookUrl: 'http://localhost:9001',
+        ...config,
+      },
+      logger,
+      performance,
+      timeItAsync,
+      outputStream: stream,
+    });
+    results = flatten(results.map(r => r.resultsOrErr));
+    for (const testResults of results) {
+      const sessionUrl = `${serverUrl}/api/sessions/batches/${encodeURIComponent(
+        testResults.getBatchId(),
+      )}/${encodeURIComponent(testResults.getId())}`;
+
+      const session = await fetch(sessionUrl).then(r => r.json());
+      expect(session.startInfo.parentBranchBaselineSavedBefore).to.be.undefined;
+    }
+  });
 });