+++ packages/sdk-coverage-tests/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 2.3.8 - 2021/2/23
+
+- fix bug with unskipped and unexecuted tests
+
 ## 2.3.7 - 2021/2/6
 
 - Include non-emitted tests in the report