+++ packages/visual-grid-client/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 15.6.0 - 2021/2/26
+
+- support `ignoreGitMergeBase`
+
 ## 15.5.22 - 2021/2/18
 
 - updated to @applitools/eyes-sdk-core@12.14.10 (from 12.14.8)